refactor: Item 도메인 리팩토링
