+++ src/main/java/com/example/tradedemo/domain/item/controller/ItemController.java
@@ -4,10 +4,9 @@ import com.example.tradedemo.common.dto.ApiResponse;
 import com.example.tradedemo.common.dto.PageResponse;
 import com.example.tradedemo.domain.item.dto.GetItemResponse;
 import com.example.tradedemo.domain.item.dto.SearchItemRequest;
-import com.example.tradedemo.domain.item.service.ItemService;
+import com.example.tradedemo.domain.item.facade.ItemFacade;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.web.PagedModel;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -19,11 +18,11 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor()
 public class ItemController {
 
-    private final ItemService itemService;
+    private final ItemFacade itemFacade;
 
     @GetMapping("/api/v1/items/{itemId}")
     public ResponseEntity<ApiResponse<GetItemResponse>> getItem(@PathVariable Long itemId) {
-        GetItemResponse res = itemService.getItem(itemId);
+        GetItemResponse res = itemFacade.getItem(itemId);
 
         return ResponseEntity.status(HttpStatus.OK)
                 .body(ApiResponse.success(String.valueOf(HttpStatus.OK.value()), res));
@@ -32,7 +31,7 @@ public class ItemController {
     @GetMapping("/api/v1/items")
     public ResponseEntity<ApiResponse<PageResponse<GetItemResponse>>> getManyItem(
             @Valid @ModelAttribute SearchItemRequest req) {
-        PageResponse<GetItemResponse> res = itemService.getManyItems(req);
+        PageResponse<GetItemResponse> res = itemFacade.getManyItems(req);
 
         return ResponseEntity.status(HttpStatus.OK)
                 .body(ApiResponse.success(String.valueOf(HttpStatus.OK.value()), res));
@@ -40,7 +39,7 @@ public class ItemController {
 
     @GetMapping("/api/v2/items/{itemId}")
     public ResponseEntity<ApiResponse<GetItemResponse>> getItemV2(@PathVariable Long itemId) {
-        GetItemResponse res = itemService.getItemV2(itemId);
+        GetItemResponse res = itemFacade.getItemV2(itemId);
 
         return ResponseEntity.status(HttpStatus.OK)
                 .body(ApiResponse.success(String.valueOf(HttpStatus.OK.value()), res));
@@ -49,7 +48,7 @@ public class ItemController {
     @GetMapping("/api/v2/items")
     public ResponseEntity<ApiResponse<PageResponse<GetItemResponse>>> getManyItemV2(
             @Valid @ModelAttribute SearchItemRequest req) {
-        PageResponse<GetItemResponse> res = itemService.getManyItemsV2(req);
+        PageResponse<GetItemResponse> res = itemFacade.getManyItemsV2(req);
 
         return ResponseEntity.status(HttpStatus.OK)
                 .body(ApiResponse.success(String.valueOf(HttpStatus.OK.value()), res));
@@ -57,7 +56,7 @@ public class ItemController {
 
     @GetMapping("/api/v3/items/{itemId}")
     public ResponseEntity<ApiResponse<GetItemResponse>> getItemV3(@PathVariable Long itemId) {
-        GetItemResponse res = itemService.getItemV3(itemId);
+        GetItemResponse res = itemFacade.getItemV3(itemId);
 
         return ResponseEntity.status(HttpStatus.OK)
                 .body(ApiResponse.success(String.valueOf(HttpStatus.OK.value()), res));
@@ -66,7 +65,7 @@ public class ItemController {
     @GetMapping("/api/v3/items")
     public ResponseEntity<ApiResponse<PageResponse<GetItemResponse>>> getManyItemV3(
             @Valid @ModelAttribute SearchItemRequest req) {
-        PageResponse<GetItemResponse> res = itemService.getManyItemsV3(req);
+        PageResponse<GetItemResponse> res = itemFacade.getManyItemsV3(req);
 
         return ResponseEntity.status(HttpStatus.OK)
                        .body(ApiResponse.success(String.valueOf(HttpStatus.OK.value()), res));
