+++ src/main/java/com/example/tradedemo/domain/item/service/ItemService.java
@@ -21,7 +21,6 @@ import com.example.tradedemo.common.exception.ErrorEnum;
 @RequiredArgsConstructor
 public class ItemService {
     private final ItemRepository itemRepository;
-    private final ItemCacheService itemCacheService;
 
     @Transactional(readOnly = true)
     public Item findItem(Long itemId) {
@@ -31,18 +30,8 @@ public class ItemService {
     }
 
     @Transactional(readOnly = true)
-    public GetItemResponse getItem(Long itemId) {
-        Item item = itemRepository.findById(itemId).orElseThrow(
-                () -> new ServiceException(ErrorEnum.ERR_ITEM_NOT_FOUND)
-        );
-
-        return GetItemResponse.of(item);
-    }
-
-    @Transactional(readOnly = true)
-    public PageResponse<GetItemResponse> getManyItems(SearchItemRequest req) {
-        Page<Item> items = itemRepository.searchItem(req);
-        return PageResponse.of(items.map(GetItemResponse::of));
+    public Page<Item> findItems(SearchItemRequest req) {
+        return itemRepository.searchItem(req);
     }
 
     @Cacheable(
@@ -51,10 +40,7 @@ public class ItemService {
     )
     @Transactional(readOnly = true)
     public GetItemResponse getItemV2(Long itemId) {
-        Item item = itemRepository.findById(itemId).orElseThrow(
-                () -> new ServiceException(ErrorEnum.ERR_ITEM_NOT_FOUND)
-        );
-
+        Item item = findItem(itemId);
         return GetItemResponse.of(item);
     }
 
@@ -67,49 +53,5 @@ public class ItemService {
     public PageResponse<GetItemResponse> getManyItemsV2(SearchItemRequest req) {
         Page<Item> items = itemRepository.searchItem(req);
         return PageResponse.of(items.map(GetItemResponse::of));
-    }
-
-    @Transactional(readOnly = true)
-    public GetItemResponse getItemV3(Long itemId) {
-        String cacheKey = itemCacheService.getItemIdCacheKey(itemId);
-
-        GetItemResponse cached = itemCacheService.getItem(cacheKey);
-        if (cached != null) {
-            return cached;
-        }
-
-        Item item = itemRepository.findById(itemId).orElseThrow(
-                () -> new ServiceException(ErrorEnum.ERR_ITEM_NOT_FOUND)
-        );
-
-        GetItemResponse res = GetItemResponse.of(item);
-        itemCacheService.setItem(cacheKey, res);
-
-        return res;
-    }
-
-    @Transactional(readOnly = true)
-    public PageResponse<GetItemResponse> getManyItemsV3(SearchItemRequest req) {
-        boolean shouldCache = itemCacheService.shouldCacheSearchItemRequest(req);
-
-        // 만약 cache를 안하기로 한 요청이라면 redis를 부르지도 않았을 것이기 때문에 DB로 바로 조회
-        if (!shouldCache) {
-            Page<Item> items = itemRepository.searchItem(req);
-            return PageResponse.of(items.map(GetItemResponse::of));
-        }
-
-        String cacheKey = itemCacheService.getSearchItemRequestCacheKey(req);
-
-        PageResponse<GetItemResponse> cached = itemCacheService.getItemList(cacheKey);
-        if (cached != null) {
-            return cached;
-        }
-
-        Page<Item> items = itemRepository.searchItem(req);
-        PageResponse<GetItemResponse> res = PageResponse.of(items.map(GetItemResponse::of));
-
-        itemCacheService.setItemList(cacheKey, res);
-
-        return res;
     }
 }
